Java programmers spend a lot of time and write a lot of bugs trying to make a BufferedReader from an OutputStream, PrintStream, or Writer.
Oftentimes, the only purpose of the stream is to pipe `String`s across an application, in which case the programmer shouldn't have to worry about `Charset` or `byte[]` at all, not to mention the performance impact of serializing and deserializing the `String`s.

Thus StringPrinter.  It takes a `Consumer<String>`, and provides `print()` and `println()` methods which make it a drop-in replacement for `PrintStream`.  Whenever a string is printed to the StringPrinter, it simply passes it to the Consumer.  StringPrinter can create any of Java's many String-stream substitutes (`OutputStream`, `PrintStream`, or `Writer`), and anything that gets pumped into them will show up (correctly decoded) in the `Consumer<String>`.