Renamed ErrorHandler to Errors.  This required a big 2.0 bump, but nobody uses Durian yet except DiffPlug, so that's okay.
After using ErrorHandler on a project with a lot of dependencies (namely Spotless, which has all of Gradle's API deps), it turns out there are a TON of classes named ErrorHandler.  Since ErrorHandler is meant to be a core utility class, it is a bummer that it conflicts with so many other names.

Errors has the advantage that it's a little shorter, and a little "fluent-y".  "Errors.suppress().run()" and "Errors.log().get()" read very nicely.  It has the downside of being an awkward name for a class which actually does have subclasses.  Whats the plural of "Errors"?  "Errorses", I guess.  That's what I used in ErrorsExample anyway.

Here are the name conflicts in the jars needed by gradleApi():
org.xml.sax.ErrorHandler
org.mortbay.jetty.handler.ErrorHandler
org.apache.log4j.spi.ErrorHandler
org.apache.jasper.compiler.ErrorHandler
and several jdk.internal and com.sun.internal classes as well